Add semi_train.sh and pseudo_label.py scripts, and update get_progress_bar function

+++ dataload/utils.py
@@ -103,8 +103,8 @@ def load_label(label_path: str, image_spacing: np.ndarray, min_d = 0, min_size =
     bboxes = np.array(info[BBOXES]) # (n, 2, 3)
     nodule_sizes = np.array(info[NODULE_SIZE]) # (n, 1)
     if len(bboxes) == 0:
-        label = {ALL_LOC: np.zeros((0, 3), dtype=np.float32),
-                ALL_CLS: np.zeros((0, 3), dtype=np.float32),
+        label = {ALL_LOC: np.zeros((0, 3), dtype=np.float64),
+                ALL_CLS: np.zeros((0, 3), dtype=np.float64),
                 ALL_RAD: np.zeros((0,), dtype=np.int32),
                 NODULE_SIZE: np.zeros((0,), dtype=np.int32)}
     else:
@@ -112,8 +112,8 @@ def load_label(label_path: str, image_spacing: np.ndarray, min_d = 0, min_size =
         if (bboxes < 0).any():
             print(f'Warning: {label_path} has negative values')
         # calculate center of bboxes
-        all_loc = ((bboxes[:, 0] + bboxes[:, 1]) / 2).astype(np.float32) # (y, x, z)
-        all_rad = (bboxes[:, 1] - bboxes[:, 0]).astype(np.float32) # (y, x, z)
+        all_loc = ((bboxes[:, 0] + bboxes[:, 1]) / 2).astype(np.float64) # (y, x, z)
+        all_rad = (bboxes[:, 1] - bboxes[:, 0]).astype(np.float64) # (y, x, z)
         
         all_loc = all_loc[:, [2, 0, 1]] # (z, y, x)
         all_rad = all_rad[:, [2, 0, 1]] # (z, y, x)
@@ -126,8 +126,8 @@ def load_label(label_path: str, image_spacing: np.ndarray, min_d = 0, min_size =
         all_cls = np.zeros((all_loc.shape[0],), dtype=np.int32)
 
         if np.sum(valid_mask) == 0:
-            label = {ALL_LOC: np.zeros((0, 3), dtype=np.float32),
-                    ALL_CLS: np.zeros((0, 3), dtype=np.float32),
+            label = {ALL_LOC: np.zeros((0, 3), dtype=np.float64),
+                    ALL_CLS: np.zeros((0, 3), dtype=np.float64),
                     ALL_RAD: np.zeros((0,), dtype=np.int32),
                     NODULE_SIZE: np.zeros((0,), dtype=np.int32)}
         else:
